move out get_reference_from_active_record_association
